requirements/base.txt
# Essential
django>=3.1.4
djangorestframework>=3.12.2
django-filter>=2.4.0
django-environ>=0.4.5
mysqlclient>=2.0.2
requests>=2.25.1
click>=7.1.2
argon2-cffi>=20.1.0
PyJWT>=2.0.0
celery==4.4.7
flower==0.9.5
Pillow>=8.0.1
django-celery-beat>=2.1.0
django-rest-multiple-models>=2.1.3
